Move presubmitFilter tests.
